Important fundamental changes to how the the following functions work:
  - VECM
  - CIModelFrame

Changing how dummies are introduced. Dummy variables are now introduced through the exogenous variable xts data argument. This simplified the code since once less input argument will be needed and dummies will be handled similarly to exogenous variables.

Exogenous variables are now used such that their timing matches the timing the endogenous variables. Before, exogenous variables would enter on the right hand side of the regression equation with a time of t. Now they enter with the same lag structure as the endogenous variables. This seems more natural to me.

Deterministic trends used to be build by specifying an integer which was a polynomial order for the highest order polynomial introduced. I do not think this is the best way to to it, for several reasons. First, it is tricky to understand when setting up the system, what exactly is done. Second, there is no theory for including higher order polynomials that a linear trend. This has been changed, so that one sets the structure using a string that can be one of the following; "none", "mean" "drift", "qtrend", "urmean".

The construction of the determinitic trend has been build into a separate function to make CIModelFrame simpler. This new funciton is called: dettrendbuild()

Seasonal dummies are introduced in the same way as before, and so is not included in the exogenous variables as other types of dummies would be.

Generally, the functions has been commented to ease reading of the code.