A fundamental change to the structure of the code.
I have split the code up in many different files based on themes, in stead of having all functions in one single file. This is done because it becomes increadible tedious to write and maintain code when having a single very long document.

Having split the functions into more files will also make it easier to to incremental work on specific parts of the library.

With this change, there are a number of other minor changes and an addition to the code given by a number of functions for prediction. These have not been finalized yet and needs some more work before they are operational.